Patch Delete method + patch bug in Home Page front when JSON.parse is…

+++ backend/src/modules/users/users.controller.ts
@@ -32,9 +32,10 @@ import {
   ApiTags,
   ApiUnauthorizedResponse,
 } from '@nestjs/swagger';
-import { UserSanitize } from 'src/interfaces/user.interfaces';
+import { JWTPayload, UserSanitize } from 'src/interfaces/user.interfaces';
 import { User } from './entities/user.entity';
 import { AuthService } from '../auth/auth.service';
+import { JwtService } from '@nestjs/jwt';
 
 @UseGuards(AuthGuard)
 @ApiTags('Users Module')
@@ -43,6 +44,7 @@ export class UsersController {
   constructor(
     private readonly usersService: UsersService,
     private readonly authService: AuthService,
+    private jwtService: JwtService,
   ) {}
 
   // Recieving a POST request to create a new user
@@ -132,9 +134,22 @@ export class UsersController {
   @ApiInternalServerErrorResponse({ description: 'Internal server error' })
   async deleteUser(
     @Param('id', ParseIntPipe) id: number,
+    @Req() req: Request,
     @Res() res: Response,
   ) {
     try {
+      const separator = req.headers.authorization?.includes('%20')
+        ? '%20'
+        : ' ';
+
+      const [type, token] = req.headers.authorization?.split(separator);
+
+      if (type != 'Bearer') return res.status(401).send();
+
+      const payload: JWTPayload = <JWTPayload>this.jwtService.decode(token);
+
+      if (id !== payload.sub) return res.status(401).send();
+
       await this.usersService.deleteUser(id);
       return res.status(HttpStatus.OK).send();
     } catch (e) {
